% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_symbol_entrez.R
\name{convert_symbol_entrez}
\alias{convert_symbol_entrez}
\title{Add together two numbers}
\usage{
convert_symbol_entrez(gene)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Add together two numbers
}
\examples{
add(1, 1)
add(10, 1)
}
