% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudobulk_score.R
\name{pseudobulk_score}
\alias{pseudobulk_score}
\title{Pseudobulk Patients into representative data point}
\usage{
pseudobulk_score(newdf)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Pseudobulk Patients into representative data point
}
\examples{
add(1, 1)
add(10, 1)
}
