% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_sig.R
\name{score_sig}
\alias{score_sig}
\title{Add together two numbers}
\usage{
score_sig(seurat)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Add together two numbers
}
\examples{
add(1, 1)
add(10, 1)
}
