% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_to_cellphonedb.R
\name{seurat_to_cellphonedb}
\alias{seurat_to_cellphonedb}
\title{Add together two numbers}
\usage{
seurat_to_cellphonedb(data_object)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Add together two numbers
}
\examples{
add(1, 1)
add(10, 1)
}
