% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_to_csv.R
\name{seurat_to_csv}
\alias{seurat_to_csv}
\title{Seurat Object to CSV Conversion Wrapper}
\usage{
seurat_to_csv(
  seurat_object,
  data_slot = "counts",
  matrix_file = "counts.csv",
  meta_file = "meta.csv"
)
}
\arguments{
\item{data_slot}{Choose which data slot to pull from. Default is (counts)}
}
\value{
csv file output into working directory
}
\description{
CSV Conversion Wrapper for Seurat Objects
}
\examples{
seurat_to_csv(pbmc.seurat)

}
